Add files via upload
# # Exploratory Data Analysis (EDA) in Pandas ##
EDA is basically the first look at your data. This process looks at patterns in the data, understanding the relationship within the features, and look at outliers within the dataset.
You are also looking for mistake and missing values that will need to be cleaned up.
The following are some of the most popular EDA techniques.